Refactor "zoom" function in MapView
Initially, I chose to use the mapview setRegion signature that allowed latitudinal and longitudinal meter properties to define the "zoom" level of the map. After reading online, it seemed like the more appropriate way to do this is to use the span (MKCoordinateSpan) property to set the map zoom level.